[Refactor] refactor config directory
